Create a bigger split between the server and cloud version of bitbucket.
This commit serves to be able to use this plugin with the Bitbucket server instance. I'm also creating an plugin for Bitbucket server for this to work.
This will send ( just like the cloud ) events when a PR is created/updated and when a push is done.
This way we can use the same functionality for both Cloud and Server.